auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic
pydantic-settings
PyJWT
httpx
